third_party/WebKit/LayoutTests/media/track/track-webvtt-tc001-utf8.html

<!DOCTYPE html>
<meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
<title>Tests that UTF-8 encoded characters are recognized properly and that different encodings (iconv) are not recognized as a WebVTT file (we do allow it, it just looks ugly).</title>
<video>
    <track src="captions-webvtt/tc001-utf8.vtt">
    <track src="captions-webvtt/tc001-iso2022jp3.vtt">
</video>
<script src="track-helpers.js"></script>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
async_test(function(t) {
    var video = document.querySelector("video");
    
    var trackElements = document.querySelectorAll("track");
    for (var i = 0; i < video.textTracks.length; i++)
        trackElements[i].onload = t.step_func(trackLoaded);

    enableAllTextTracks(video.textTracks);

    var numberOfTracksLoaded = 0;
    function trackLoaded() {
        numberOfTracksLoaded++;
        if (numberOfTracksLoaded != 2)
            return;

        testTrack0();
        testTrack1();
        t.done();
    }

    function testTrack0() {
        var expected = [
            {
                id : "1",
                startTime : 0,
                endTime : 30.5,
                text : "景気判断"
            },
            {
                id : "2",
                startTime : 31,
                endTime : 1200.5,
                text : "電力不足"
            }
        ];

        var cues = video.textTracks[0].cues;
        assert_equals(cues.length, 2);
        assert_cues_equal(cues, expected);
    }

    function testTrack1() {
        assert_equals(video.textTracks[1].cues.length, 2);
    }
});
</script>
